+++ storefront/src/modules/order/components/shipping-details/index.tsx
@@ -27,15 +27,12 @@ const ShippingDetails = ({ order }: ShippingDetailsProps) => {
             {order.shipping_address?.last_name}
           </Text>
           <Text className="txt-medium text-ui-fg-subtle">
-            {order.shipping_address?.address_1}{" "}
-            {order.shipping_address?.address_2}
+            {order.shipping_address?.address_1}
+            {order.shipping_address?.address_2 &&
+              `, ${order.shipping_address.address_2}`}
           </Text>
           <Text className="txt-medium text-ui-fg-subtle">
-            {order.shipping_address?.postal_code},{" "}
-            {order.shipping_address?.city}
-          </Text>
-          <Text className="txt-medium text-ui-fg-subtle">
-            {order.shipping_address?.country_code?.toUpperCase()}
+            {order.shipping_address?.city}, {order.shipping_address?.province}
           </Text>
         </div>